requirements.txt
Django==1.8.5
heroku==0.1.4
python-dateutil==1.5
requests==2.8.1
wheel==0.24.0
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
certifi==2015.4.28
openpyxl==2.3.0
whitenoise
django-bootstrap3
django-dajax
django-dajaxice==0.7

